+++ changelog.md
@@ -1,3 +1,8 @@
-# 0.0.1
+# Changelog
 
+# 0.0.2
+- changed to consuming codemirror from bower
+- we now concatenate all modes into the main codemirror.js file
+
+# 0.0.1
 - Using codemirror 4.4.0 (http://codemirror.net/doc/compress.html)